refactor: extract app prompt helpers
